requirements.txt
ddt==1.7.2
dimod==0.12.14
docplex==2.25.236
dwave_cloud_client==0.11.3
dwave_drivers==0.4.4
dwave_inspector==0.4.4
dwave_inspectorapp==0.3.1
dwave_preprocessing==0.6.5
dwave_samplers==1.2.0
dwave_system==1.23.0
ipython==8.12.3
nlopt==2.7.1
numpy==1.23.5
ortools==9.10.4067
pandas==2.2.2
plotly==5.19.0
pyenchant==3.2.2
python-dotenv==1.0.1
qiskit==1.0.2
qiskit_ibm_runtime==0.23.0
qiskit_ibmq_provider==0.20.2
qiskit_optimization==0.6.1
rustworkx==0.14.2
scipy==1.13.1
setuptools==70.0.0
SQSnobFit==0.4.5
